Refactoring VIIRS/ATMS SDR reading, making a common base class, and fix tests
