+++ satpy/readers/viirs_atms_sdr_utils.py
@@ -18,35 +18,292 @@
 
 """Common utilities for reading VIIRS and ATMS SDR data."""
 
+import logging
+from datetime import datetime, timedelta
 
-DATASET_KEYS = {'GDNBO': 'VIIRS-DNB-GEO',
-                'SVDNB': 'VIIRS-DNB-SDR',
-                'GITCO': 'VIIRS-IMG-GEO-TC',
-                'GIMGO': 'VIIRS-IMG-GEO',
-                'SVI01': 'VIIRS-I1-SDR',
-                'SVI02': 'VIIRS-I2-SDR',
-                'SVI03': 'VIIRS-I3-SDR',
-                'SVI04': 'VIIRS-I4-SDR',
-                'SVI05': 'VIIRS-I5-SDR',
-                'GMTCO': 'VIIRS-MOD-GEO-TC',
-                'GMODO': 'VIIRS-MOD-GEO',
-                'SVM01': 'VIIRS-M1-SDR',
-                'SVM02': 'VIIRS-M2-SDR',
-                'SVM03': 'VIIRS-M3-SDR',
-                'SVM04': 'VIIRS-M4-SDR',
-                'SVM05': 'VIIRS-M5-SDR',
-                'SVM06': 'VIIRS-M6-SDR',
-                'SVM07': 'VIIRS-M7-SDR',
-                'SVM08': 'VIIRS-M8-SDR',
-                'SVM09': 'VIIRS-M9-SDR',
-                'SVM10': 'VIIRS-M10-SDR',
-                'SVM11': 'VIIRS-M11-SDR',
-                'SVM12': 'VIIRS-M12-SDR',
-                'SVM13': 'VIIRS-M13-SDR',
-                'SVM14': 'VIIRS-M14-SDR',
-                'SVM15': 'VIIRS-M15-SDR',
-                'SVM16': 'VIIRS-M16-SDR',
-                'IVCDB': 'VIIRS-DualGain-Cal-IP',
-                'SATMS': 'ATMS-SDR',
-                'GATMO': 'ATMS-SDR-GEO',
-                'TATMS': 'ATMS-TDR'}
+import dask.array as da
+import numpy as np
+import xarray as xr
+
+from satpy.readers.hdf5_utils import HDF5FileHandler
+
+NO_DATE = datetime(1958, 1, 1)
+EPSILON_TIME = timedelta(days=2)
+LOG = logging.getLogger(__name__)
+
+
+VIIRS_DATASET_KEYS = {'GDNBO': 'VIIRS-DNB-GEO',
+                      'SVDNB': 'VIIRS-DNB-SDR',
+                      'GITCO': 'VIIRS-IMG-GEO-TC',
+                      'GIMGO': 'VIIRS-IMG-GEO',
+                      'SVI01': 'VIIRS-I1-SDR',
+                      'SVI02': 'VIIRS-I2-SDR',
+                      'SVI03': 'VIIRS-I3-SDR',
+                      'SVI04': 'VIIRS-I4-SDR',
+                      'SVI05': 'VIIRS-I5-SDR',
+                      'GMTCO': 'VIIRS-MOD-GEO-TC',
+                      'GMODO': 'VIIRS-MOD-GEO',
+                      'SVM01': 'VIIRS-M1-SDR',
+                      'SVM02': 'VIIRS-M2-SDR',
+                      'SVM03': 'VIIRS-M3-SDR',
+                      'SVM04': 'VIIRS-M4-SDR',
+                      'SVM05': 'VIIRS-M5-SDR',
+                      'SVM06': 'VIIRS-M6-SDR',
+                      'SVM07': 'VIIRS-M7-SDR',
+                      'SVM08': 'VIIRS-M8-SDR',
+                      'SVM09': 'VIIRS-M9-SDR',
+                      'SVM10': 'VIIRS-M10-SDR',
+                      'SVM11': 'VIIRS-M11-SDR',
+                      'SVM12': 'VIIRS-M12-SDR',
+                      'SVM13': 'VIIRS-M13-SDR',
+                      'SVM14': 'VIIRS-M14-SDR',
+                      'SVM15': 'VIIRS-M15-SDR',
+                      'SVM16': 'VIIRS-M16-SDR',
+                      'IVCDB': 'VIIRS-DualGain-Cal-IP'}
+ATMS_DATASET_KEYS = {'SATMS': 'ATMS-SDR',
+                     'GATMO': 'ATMS-SDR-GEO',
+                     'TATMS': 'ATMS-TDR'}
+
+DATASET_KEYS = {}
+DATASET_KEYS.update(VIIRS_DATASET_KEYS)
+DATASET_KEYS.update(ATMS_DATASET_KEYS)
+
+
+def get_file_units(dataset_id, ds_info):
+    """Get file units from metadata."""
+    file_units = ds_info.get("file_units")
+    if file_units is None:
+        LOG.debug("Unknown units for file key '%s'", dataset_id)
+    return file_units
+
+
+class JPSS_SDR_FileHandler(HDF5FileHandler):
+    """Base class for reading JPSS VIIRS & ATMS SDR HDF5 Files."""
+
+    def __init__(self, filename, filename_info, filetype_info, use_tc=None, **kwargs):
+        """Initialize file handler."""
+        super().__init__(filename, filename_info, filetype_info, **kwargs)
+
+    def _parse_datetime(self, datestr, timestr):
+        try:
+            datetime_str = datestr + timestr
+        except TypeError:
+            datetime_str = str(datestr.astype(str)) + str(timestr.astype(str))
+        time_val = datetime.strptime(datetime_str, '%Y%m%d%H%M%S.%fZ')
+        if abs(time_val - NO_DATE) < EPSILON_TIME:
+            # catch rare case when SDR files have incorrect date
+            raise ValueError("Datetime invalid {}".format(time_val))
+        return time_val
+
+    @property
+    def start_time(self):
+        """Get start time."""
+        dataset_group = DATASET_KEYS[self.datasets[0]]
+        default_start_date = 'Data_Products/{dataset_group}/{dataset_group}_Aggr/attr/AggregateBeginningDate'
+        default_start_time = 'Data_Products/{dataset_group}/{dataset_group}_Aggr/attr/AggregateBeginningTime'
+        date_var_path = self.filetype_info.get('start_date', default_start_date).format(dataset_group=dataset_group)
+        time_var_path = self.filetype_info.get('start_time', default_start_time).format(dataset_group=dataset_group)
+        return self._parse_datetime(self[date_var_path], self[time_var_path])
+
+    @property
+    def end_time(self):
+        """Get end time."""
+        dataset_group = DATASET_KEYS[self.datasets[0]]
+        default_end_date = 'Data_Products/{dataset_group}/{dataset_group}_Aggr/attr/AggregateEndingDate'
+        default_end_time = 'Data_Products/{dataset_group}/{dataset_group}_Aggr/attr/AggregateEndingTime'
+        date_var_path = self.filetype_info.get('end_date', default_end_date).format(dataset_group=dataset_group)
+        time_var_path = self.filetype_info.get('end_time', default_end_time).format(dataset_group=dataset_group)
+        return self._parse_datetime(self[date_var_path], self[time_var_path])
+
+    @property
+    def start_orbit_number(self):
+        """Get start orbit number."""
+        dataset_group = DATASET_KEYS[self.datasets[0]]
+        default = 'Data_Products/{dataset_group}/{dataset_group}_Aggr/attr/AggregateBeginningOrbitNumber'
+        start_orbit_path = self.filetype_info.get('start_orbit', default).format(dataset_group=dataset_group)
+        return int(self[start_orbit_path])
+
+    @property
+    def end_orbit_number(self):
+        """Get end orbit number."""
+        dataset_group = DATASET_KEYS[self.datasets[0]]
+        default = 'Data_Products/{dataset_group}/{dataset_group}_Aggr/attr/AggregateEndingOrbitNumber'
+        end_orbit_path = self.filetype_info.get('end_orbit', default).format(dataset_group=dataset_group)
+        return int(self[end_orbit_path])
+
+    @property
+    def platform_name(self):
+        """Get platform name."""
+        default = '/attr/Platform_Short_Name'
+        platform_path = self.filetype_info.get(
+            'platform_name', default).format(**self.filetype_info)
+        platform_dict = {'NPP': 'Suomi-NPP',
+                         'JPSS-1': 'NOAA-20',
+                         'J01': 'NOAA-20',
+                         'JPSS-2': 'NOAA-21',
+                         'J02': 'NOAA-21'}
+        return platform_dict.get(self[platform_path], self[platform_path])
+
+    @property
+    def sensor_name(self):
+        """Get sensor name."""
+        dataset_group = DATASET_KEYS[self.datasets[0]]
+        default = 'Data_Products/{dataset_group}/attr/Instrument_Short_Name'
+        sensor_path = self.filetype_info.get(
+            'sensor_name', default).format(dataset_group=dataset_group)
+        return self[sensor_path].lower()
+
+    def scale_swath_data(self, data, scaling_factors, dataset_group):
+        """Scale swath data using scaling factors and offsets.
+
+        Multi-granule (a.k.a. aggregated) files will have more than the usual two values.
+        """
+        rows_per_gran = self._get_rows_per_granule(dataset_group)
+        factors = self._mask_and_reshape_factors(scaling_factors)
+        data = self._map_and_apply_factors(data, factors, rows_per_gran)
+        return data
+
+    @staticmethod
+    def _mask_and_reshape_factors(factors):
+        factors = factors.where(factors > -999, np.float32(np.nan))
+        return factors.data.reshape((-1, 2)).rechunk((1, 2))  # make it so map_blocks happens per factor
+
+    @staticmethod
+    def _map_and_apply_factors(data, factors, rows_per_gran):
+        # The user may have requested a different chunking scheme, but we need
+        # per granule chunking right now so factor chunks map 1:1 to data chunks
+        old_chunks = data.chunks
+        dask_data = data.data.rechunk((tuple(rows_per_gran), data.data.chunks[1]))
+        dask_data = da.map_blocks(_apply_factors, dask_data, factors,
+                                  chunks=dask_data.chunks, dtype=data.dtype,
+                                  meta=np.array([[]], dtype=data.dtype))
+        data = xr.DataArray(dask_data.rechunk(old_chunks),
+                            dims=data.dims, coords=data.coords,
+                            attrs=data.attrs)
+        return data
+
+    @staticmethod
+    def _scale_factors_for_units(factors, file_units, output_units):
+        if file_units == "W cm-2 sr-1" and output_units == "W m-2 sr-1":
+            LOG.debug("Adjusting scaling factors to convert '%s' to '%s'",
+                      file_units, output_units)
+            factors = factors * 10000.
+        elif file_units == "1" and output_units == "%":
+            LOG.debug("Adjusting scaling factors to convert '%s' to '%s'",
+                      file_units, output_units)
+            factors = factors * 100.
+        else:
+            raise ValueError("Don't know how to convert '{}' to '{}'".format(
+                file_units, output_units))
+        return factors
+
+    @staticmethod
+    def _get_valid_scaling_factors(factors):
+        if factors is None:
+            factors = np.array([1, 0], dtype=np.float32)
+            factors = xr.DataArray(da.from_array(factors, chunks=1))
+        else:
+            factors = factors.where(factors != -999., np.float32(np.nan))
+        return factors
+
+    def _adjust_scaling_factors(self, factors, file_units, output_units):
+        """Adjust scaling factors ."""
+        if file_units == output_units:
+            LOG.debug("File units and output units are the same (%s)", file_units)
+            return factors
+        factors = self._get_valid_scaling_factors(factors)
+        return self._scale_factors_for_units(factors, file_units, output_units)
+
+    def _get_scaling_factors(self, file_units, output_units, factor_var_path):
+        """Get file scaling factors and scale according to expected units."""
+        factors = self.get(factor_var_path)
+        factors = self._adjust_scaling_factors(factors, file_units, output_units)
+        return factors
+
+    @staticmethod
+    def expand_single_values(var, scans):
+        """Expand single valued variable to full scan lengths."""
+        if scans.size == 1:
+            return var
+        else:
+            expanded = np.repeat(var, scans)
+            expanded.attrs = var.attrs
+            expanded.rename({expanded.dims[0]: 'y'})
+            return expanded
+
+    def _scan_size(self, dataset_group_name):
+        """Get how many rows of data constitute one scanline."""
+        if 'ATM' in dataset_group_name:
+            scan_size = 1
+        elif 'I' in dataset_group_name:
+            scan_size = 32
+        else:
+            scan_size = 16
+        return scan_size
+
+    def concatenate_dataset(self, dataset_group, var_path):
+        """Concatenate dataset."""
+        scan_size = self._scan_size(dataset_group)
+        scans = self._get_scans_per_granule(dataset_group)
+        start_scan = 0
+        data_chunks = []
+        scans = xr.DataArray(scans)
+        variable = self[var_path]
+        # check if these are single per-granule value
+        if variable.size != scans.size:
+            for gscans in scans.values:
+                data_chunks.append(self[var_path].isel(y=slice(start_scan, start_scan + gscans * scan_size)))
+                start_scan += gscans * scan_size
+            return xr.concat(data_chunks, 'y')
+        else:
+            return self.expand_single_values(variable, scans)
+
+    def _get_rows_per_granule(self, dataset_group):
+        scan_size = self._scan_size(dataset_group)
+        scans_per_gran = self._get_scans_per_granule(dataset_group)
+        return [scan_size * gran_scans for gran_scans in scans_per_gran]
+
+    def _get_scans_per_granule(self, dataset_group):
+        number_of_granules_path = 'Data_Products/{dataset_group}/{dataset_group}_Aggr/attr/AggregateNumberGranules'
+        nb_granules_path = number_of_granules_path.format(dataset_group=DATASET_KEYS[dataset_group])
+        scans = []
+        for granule in range(self[nb_granules_path]):
+            scans_path = 'Data_Products/{dataset_group}/{dataset_group}_Gran_{granule}/attr/N_Number_Of_Scans'
+            scans_path = scans_path.format(dataset_group=DATASET_KEYS[dataset_group], granule=granule)
+            scans.append(self[scans_path])
+        return scans
+
+    def mask_fill_values(self, data, ds_info):
+        """Mask fill values."""
+        is_floating = np.issubdtype(data.dtype, np.floating)
+
+        if is_floating:
+            # If the data is a float then we mask everything <= -999.0
+            fill_max = np.float32(ds_info.pop("fill_max_float", -999.0))
+            return data.where(data > fill_max, np.float32(np.nan))
+        else:
+            # If the data is an integer then we mask everything >= fill_min_int
+            fill_min = int(ds_info.pop("fill_min_int", 65528))
+            return data.where(data < fill_min, np.float32(np.nan))
+
+    def available_datasets(self, configured_datasets=None):
+        """Generate dataset info and their availablity.
+
+        See
+        :meth:`satpy.readers.file_handlers.BaseFileHandler.available_datasets`
+        for details.
+
+        """
+        for is_avail, ds_info in (configured_datasets or []):
+            if is_avail is not None:
+                yield is_avail, ds_info
+                continue
+            dataset_group = [ds_group for ds_group in ds_info['dataset_groups'] if ds_group in self.datasets]
+            if dataset_group:
+                yield True, ds_info
+            elif is_avail is None:
+                yield is_avail, ds_info
+
+
+def _apply_factors(data, factor_set):
+    return data * factor_set[0, 0] + factor_set[0, 1]
